tools/requirements.txt
# Panda Council Orchestrator Dependencies

google-generativeai>=0.3.0
anthropic>=0.18.0
